+++ app/Enums/RoomUserRole.php
@@ -13,4 +13,6 @@ final class RoomUserRole extends Enum
     const GUEST     =   0;
     const USER      =   1;
     const MODERATOR =   2;
+    const CO_OWNER  =   3;
+    const OWNER     =   4;
 }