+++ database/seeds/RolesAndPermissionsSeeder.php
@@ -19,7 +19,8 @@ class RolesAndPermissionsSeeder extends Seeder
 
         $adminPermissions[] = Permission::firstOrCreate([ 'name' => 'rooms.create' ])->id;
         $adminPermissions[] = Permission::firstOrCreate([ 'name' => 'rooms.viewAll' ])->id;
-
+        $adminPermissions[] = Permission::firstOrCreate([ 'name' => 'rooms.manage' ])->id;
+        
         $adminPermissions[] = Permission::firstOrCreate([ 'name' => 'meetings.viewAny' ])->id;
 
         $adminPermissions[] = Permission::firstOrCreate([ 'name' => 'settings.manage' ])->id;
@@ -68,6 +69,7 @@ class RolesAndPermissionsSeeder extends Seeder
 
         // Setup permission inheritances
         /// Eg. If you have permission x, you also get the permissions a,b,c
+        Permission::SetupIncludedPermissions('rooms.manage', ['rooms.create','rooms.viewAll']);
 
         Permission::SetupIncludedPermissions('meetings.viewAny', ['rooms.viewAll']);